Refactor Authark to use Modelark.
This is a big change since is the first time that an
external library is used in the application layer, though,
this library has been developed by Knowark as well.

The reusability and standardization benefits that this
refactoring brings are noteworthy and thus, this change
should be replayed in other Knowark microservices.